% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_functions.R
\name{calibration_backtest}
\alias{calibration_backtest}
\title{Conditional Calibration Backtest}
\usage{
calibration_backtest(r, q, e, s = NULL, alpha, hommel = TRUE)
}
\arguments{
\item{r}{A vector of returns}

\item{q}{A vector of Value-at-Risk forecasts}

\item{e}{A vector of expected shortfall forecasts}

\item{s}{A vector of volatility forecasts}

\item{alpha}{Scalar probability level in (0, 1)}

\item{hommel}{If TRUE, use Hommels correction, else use the classical Bonferroni correction.}
}
\value{
Returns a 2x2 matrix with p-values
}
\description{
The simple and general conditional calibration backtests of Nolde & Ziegel (2017).
}
\examples{
data(risk_forecasts)
r <- risk_forecasts$r
q <- risk_forecasts$q
e <- risk_forecasts$e
s <- risk_forecasts$s
calibration_backtest(r = r, q = q, e = e, s = s, alpha = 0.025)
}
\references{
\href{https://arxiv.org/abs/1608.05498}{Nolde & Ziegel (2007)}
}
