% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_functions.R
\name{check_inputs}
\alias{check_inputs}
\title{Check Inputs}
\usage{
check_inputs(r, q = NULL, e, s = NULL, alpha = NULL)
}
\arguments{
\item{r}{A vector of returns}

\item{q}{A vector of Value-at-Risk forecasts}

\item{e}{A vector of expected shortfall forecasts}

\item{s}{A vector of volatility forecasts}

\item{alpha}{Scalar probability level in (0, 1)}
}
\description{
Checks the input data.
}
\keyword{internal}
