% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_functions.R
\name{de_backtest}
\alias{de_backtest}
\title{Cumulative Violation Backtests}
\usage{
de_backtest(H, alpha, m = 5)
}
\arguments{
\item{H}{A vector of forecasts of cumulative violations}

\item{alpha}{Scalar probability level in (0, 1)}

\item{m}{Number of lags for the conditional test}
}
\value{
Returns a list with the following components:
\itemize{
\item pvalue_unconditional
\item pvalue_conditional
}
}
\description{
Expected shortfall backtests based on the cummulative violation series.
}
\examples{
data(risk_forecasts)
H <- risk_forecasts$H
de_backtest(H = H, alpha = 0.025)
}
\references{
\href{https://doi.org/10.1287/mnsc.2015.2342}{Du & Escanciano (2016)},
\href{https://doi.org/10.21314/JRMV.2015.131}{Costanzino & Curran (2015)}
}
