% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_functions.R
\name{er_backtest}
\alias{er_backtest}
\title{Exceedance Residuals Backtest}
\usage{
er_backtest(r, q, e, s = NULL, B = 1000)
}
\arguments{
\item{r}{A vector of returns}

\item{q}{A vector of Value-at-Risk forecasts}

\item{e}{A vector of expected shortfall forecasts}

\item{s}{A vector of volatility forecasts}

\item{B}{Number of bootstrap iterations}
}
\value{
Returns a 2x2 matrix with p-values
}
\description{
Tests whether the mean of the exceedance residuals, respectively the
mean of the standardized exceedance residuals is zero.
}
\examples{
data(df)
er_backtest(r = df$r, q = df$q1, e = df$e1, s = df$s1)
er_backtest(r = df$r, q = df$q2, e = df$e2)
}
\references{
\href{https://doi.org/10.1016/S0927-5398(00)00012-8}{McNeil & Frey (2000)}
}
