% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esback.R
\docType{package}
\name{esback}
\alias{esback}
\alias{esback-package}
\title{esback: A package for evaluating and backtesting expected shortfall forecasts}
\description{
The esback package contains functions for evaluating and backtesting
expected shortfall forecasts.
}
\section{Available backtest functions}{

\itemize{
  \item \link[=er_backtest]{Exceedance Residuals Backtest (McNeil & Frey, 2000)}
  \item \link[=calibration_backtest]{Conditional Calibration Backtest (Nolde & Ziegel, 2017)}
  \item \link[=esr_backtest]{Expeced Shortfall Regression Backtests (Bayer & Dimitriadis, 2017)}
}
}

