% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_functions.R
\name{esr_backtest}
\alias{esr_backtest}
\title{Expected Shortfall Regression Backtest}
\usage{
esr_backtest(r, e, alpha, B = 0)
}
\arguments{
\item{r}{A vector of returnss}

\item{e}{A vector of Value-at-Risk forecasts}

\item{alpha}{Scalar probability level in (0, 1)}

\item{B}{Number of bootstrap samples. Set to 0 to disable bootstrapping.}
}
\value{
Returns a 2-dim. vector with p-values
}
\description{
Regresses the expected shortfall forecasts and an intercept term on the returns
and tests the coefficients for (0, 1).
}
\examples{
data(risk_forecasts)
r <- risk_forecasts$r
e <- risk_forecasts$e
esr_backtest(r = r, e = e, alpha = 0.025)
}
\references{
Bayer & Dimitriadis (2017)
}
\seealso{
\code{\link{esr_backtest_intercept}} for the intercept ESR backtest
}
