% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_functions.R
\name{esr_backtest_intercept}
\alias{esr_backtest_intercept}
\title{Expected Shortfall Regression Intercept Backtest}
\usage{
esr_backtest_intercept(r, e, alpha, B = 0)
}
\arguments{
\item{r}{A vector of returnss}

\item{e}{A vector of Value-at-Risk forecasts}

\item{alpha}{Scalar probability level in (0, 1)}

\item{B}{Number of bootstrap samples. Set to 0 to disable bootstrapping.}
}
\value{
Returns a 2x2 matrix with p-values
}
\description{
Tests whether the expected shortfall of the forecast error r - e is zero.
}
\examples{
data(risk_forecasts)
r <- risk_forecasts$r
e <- risk_forecasts$e
esr_backtest_intercept(r = r, e = e, alpha = 0.025)
}
\references{
Bayer & Dimitriadis (2017)
}
\seealso{
\code{\link{esr_backtest}} for the bivariate ESR backtest
}
