% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{joint_loss}
\alias{joint_loss}
\title{Joint Quantile and Expected Shortfall Loss Function}
\usage{
joint_loss(r, q, e, alpha, g1 = 2L, g2 = 1L, return_mean = TRUE)
}
\arguments{
\item{r}{A vector of returnss}

\item{q}{A vector of expected shortfall forecasts}

\item{e}{A vector of Value-at-Risk forecasts}

\item{alpha}{Scalar probability level in (0, 1)}

\item{g1}{1, 2}

\item{g2}{1, 2, 3, 4, 5}

\item{return_mean}{If TRUE returns the average loss, else the individual values}
}
\description{
Computes the joint (VaR, ES) loss. Defaults to the 0-homogeneous variant.
}
\examples{
data(risk_forecasts)
joint_loss(r=risk_forecasts$r, q=risk_forecasts$q, e=risk_forecasts$e, alpha=0.025)
}
\references{
\href{http://dx.doi.org/10.1214/16-AOS1439}{Fissler & Ziegel (2016)}
}
