% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parameter_definition}
\alias{parameter_definition}
\title{Parameter Definitions}
\usage{
parameter_definition(r, q, e, s, alpha)
}
\arguments{
\item{r}{A vector of returns.}

\item{q}{A vector of Value-at-Risk forecasts.}

\item{e}{A vector of Expected Shortfall forecasts.}

\item{s}{A vector of volatility forecasts.}

\item{alpha}{Scalar probability level in (0, 1).}

\item{B}{Number of bootstrap samples. Set to 0 to disable bootstrapping.}
}
\description{
Defines commonly used parameters.
}
\keyword{internal}
