% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esback.R
\docType{package}
\name{esback}
\alias{esback}
\title{esback: A package for backtesting expected shortfall forecasts}
\description{
The esback package contains functions for backtesting expected shortfall forecasts.
}
\section{Available backtest functions}{

\itemize{
\item \link[=er_backtest]{Exceedance Residuals Backtest (McNeil & Frey, 2000)}
\item \link[=cc_backtest]{Conditional Calibration Backtest (Nolde & Ziegel, 2017)}
\item \link[=esr_backtest]{Expeced Shortfall Regression Backtests (Bayer & Dimitriadis, 2018)}
}
}

