% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_functions.R
\name{esr_backtest}
\alias{esr_backtest}
\title{Expected Shortfall Regression Backtest}
\usage{
esr_backtest(r, e, alpha)
}
\arguments{
\item{r}{A vector of returns}

\item{e}{A vector of expected shortfall forecasts}

\item{alpha}{Scalar probability level in (0, 1)}
}
\value{
Returns a 2x2 matrix with p-values
}
\description{
Contains two backtests using the esreg package.
The first is an intercept-only backtest. The forecast error,
r - e is regressed on an intercept, which is tested for zero.
The second test regresses the expected shortfall forecasts
and an intercept term on the returns and tests the coefficients for (0, 1).
}
\examples{
data(df)
esr_backtest(r = df$r, e = df$e1, alpha = 0.025)
esr_backtest(r = df$r, e = df$e2, alpha = 0.025)
}
\references{
Bayer & Dimitriadis (2017)
}
