% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_functions.R
\name{esr_backtest}
\alias{esr_backtest}
\title{Expected Shortfall Regression Backtest}
\usage{
esr_backtest(r, q, e, alpha, version, B = 0)
}
\arguments{
\item{r}{A vector of returnss}

\item{q}{A vector of expected shortfall forecasts}

\item{e}{A vector of Value-at-Risk forecasts}

\item{alpha}{Scalar probability level in (0, 1)}

\item{version}{Version of the backtest to be used}

\item{B}{Number of bootstrap samples. Set to 0 to disable bootstrapping.}
}
\value{
Returns a list with the following components:
\itemize{
\item p_value_two_sided_asymptotic
\item p_value_one_sided_asymptotic
\item p_value_two_sided_bootstrap
\item p_value_one_sided_bootstrap
}
}
\description{
This function implements multiple expected shortfall regression (esreg)
based backtests.
Using the \code{version} argument, the following backtests are available:
\enumerate{
\item Regresses the expected shortfall forecasts on
the returns and tests the ES coefficients for (0, 1).
\item Regresses the quantile and the expected shortfall forecasts on
the returns and tests the ES coefficients for (0, 1).
\item Regresses the quantile and the expected shortfall forecasts on
the returns and tests the coefficients for (0, 1, 0, 1).
\item Tests whether the expected shortfall of the forecast error r - e is zero.
\item Tests whether the expected shortfall of the forecast error 1 - r/e is zero.
}
}
\examples{
data(risk_forecasts)
r <- risk_forecasts$r
q <- risk_forecasts$q
e <- risk_forecasts$e
esr_backtest(r = r, q = q, e = e, alpha = 0.025, version = 3)
}
\references{
\href{https://arxiv.org/abs/1801.04112}{Bayer & Dimitriadis (2018)}
}
