% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{risk_forecasts}
\alias{risk_forecasts}
\title{Returns and risk forecasts for the S&P 500}
\format{A data frame with 4396 rows and 6 variables}
\usage{
data(risk_forecasts)
}
\description{
A dataset containing the returns and risk forecasts
for the S&P 500 daily log returns. All risk forecasts are for the
probability 2.5\%.
}
\details{
\itemize{
  \item r. Daily log returns from January 3, 2000 to August 15, 2017 (4396 days).
  \item q1. Value-at-Risk forecasts of the APARCH-t model.
  \item q2. Value-at-Risk forecasts of the Historical Simulation approach.
  \item e1. Expected shortfall forecasts of the APARCH-t model.
  \item e2. Expected shortfall forecasts of the Historical Simulation approach.
  \item s1. Volatility forecasts of the APARCH-t model.
}
}
\keyword{datasets}
