% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{risk_forecasts}
\alias{risk_forecasts}
\title{Returns and risk forecasts for the S&P 500 index}
\format{
A data.frame with 4396 rows and 4 variables
}
\usage{
data(risk_forecasts)
}
\description{
A dataset containing the daily log returns and risk forecasts for the S&P 500 index.
The quantile and expected shortfall forecasts are for the probability level 2.5\%.
}
\details{
Description of the variables:
\describe{
  \item{r}{Daily log returns from January 3, 2000 to September 29, 2017 (4465 days)}
  \item{q}{Value-at-Risk forecasts of the Historical Simulation approach}
  \item{e}{Expected shortfall forecasts of the Historical Simulation approach}
  \item{s}{Volatility forecasts of the Historical Simulation approach}
}
}
\keyword{datasets}
